pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "vulnscanner-ai"
version = "1.0.0"
description = "ML-powered vulnerability scanner for OWASP Top 10 (2025)"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
authors = [{ name = "VulnScanner AI Contributors" }]
keywords = ["security", "vulnerability", "owasp", "ml", "appsec", "devsecops"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Security",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "typer[all]>=0.12",
    "rich>=13.7",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "scikit-learn>=1.4",
    "xgboost>=2.0",
    "pandas>=2.2",
    "numpy>=1.26",
    "plotly>=5.20",
    "jinja2>=3.1",
    "requests>=2.31",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "python-dotenv>=1.0",
    "gitpython>=3.1",
    "libcst>=1.3",
    "shap>=0.45",
    "joblib>=1.3",
    "packaging>=24.0",
    "toml>=0.10",
    "PyYAML>=6.0",
    "weasyprint>=62.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-cov>=5.0", "black>=24.0", "ruff>=0.4", "mypy>=1.9"]
transformers = ["sentence-transformers>=3.0", "torch>=2.2"]

[project.scripts]
vulnscanner = "vulnscanner.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["vulnscanner*"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
